requirements.txt
scikit-learn
plotly
kaleido
ipykernel
numpy
matplotlib
pandas
tqdm
nbformat
tensorboard
torch
torchvision
torchaudio
wandb
MulticoreTSNE
optuna
hydra-core
python-dotenv
lightgbm
synthcity
